requirements.txt
streamlit
streamlit-searchbox
pandas
numpy
plotly
yfinance
requests
tavily-python
openai
python-dotenv